migrate three dialogs as a children component & import in users-list
