+++ packages/playground/src/components/user-management/form-user-reset-password.tsx
@@ -14,53 +14,51 @@ import {
 } from '@harnessio/canary'
 
 interface UsersProps {
-  admin: boolean
   uid: string
   display_name?: string // Add a default value of undefined
 }
 
-interface FormRemoveUserDialogProps {
+interface FormResetPasswordrDialogProps {
   user: UsersProps | null
   onClose: () => void
 }
 
-//Form Remove Admin Dialog
-export const FormRemoveUserDialog: React.FC<FormRemoveUserDialogProps> = ({ user, onClose }) => {
-  const [isRemoving, setIsRemoving] = useState(false)
-  const [removeSuccess, setRemoveSuccess] = useState(false)
+export const FormResetPasswordDialog: React.FC<FormResetPasswordrDialogProps> = ({ user, onClose }) => {
+  const [isResetting, setIsResetting] = useState(false)
+  const [resetSuccess, setResetSuccess] = useState(false)
   // Delete project handler
-  const handleRemove = () => {
-    setIsRemoving(true)
+  const handleReset = () => {
+    setIsResetting(true)
     setTimeout(() => {
-      setIsRemoving(false)
-      setRemoveSuccess(true) // Mark deletion as successful
+      setIsResetting(false)
+      setResetSuccess(true) // Mark deletion as successful
       setTimeout(() => {
         onClose() // Close the dialog
       }, 2000)
     }, 2000)
   }
-
   return (
     <AlertDialog open={true} onOpenChange={onClose}>
       <AlertDialogTrigger asChild></AlertDialogTrigger>
       <AlertDialogContent>
         <AlertDialogHeader>
-          <AlertDialogTitle>Are you sure you want to remove {user?.display_name} as an admin?</AlertDialogTitle>
+          <AlertDialogTitle>Are you sure you want to reset password for {user?.display_name}?</AlertDialogTitle>
           <AlertDialogDescription>
-            This will permanently remove a admin tag for {user?.display_name} ({user?.uid}).
+            This will send a password reset email to {user?.display_name} ({user?.uid}).
           </AlertDialogDescription>
         </AlertDialogHeader>
         <Spacer size={3} />
         <AlertDialogFooter>
-          {!isRemoving && !removeSuccess && <AlertDialogCancel onClick={onClose}>Cancel</AlertDialogCancel>}
-          {removeSuccess ? (
+          {!isResetting && !resetSuccess && <AlertDialogCancel onClick={onClose}>Cancel</AlertDialogCancel>}
+          {/* TODO: check the page flow of reset password */}
+          {resetSuccess ? (
             <Button size="default" theme="success" className="self-start pointer-events-none">
-              Users removed&nbsp;&nbsp;
+              Password reset email sent&nbsp;&nbsp;
               <Icon name="tick" size={14} />
             </Button>
           ) : (
-            <Button size="default" theme="error" className="self-start" onClick={handleRemove}>
-              {isRemoving ? 'Removing Member...' : 'Yes, remove Member'}
+            <Button size="default" theme="error" className="self-start" onClick={handleReset}>
+              {isResetting ? 'Resetting Password...' : 'Yes, reset Password'}
             </Button>
           )}
         </AlertDialogFooter>
